remove translation operator
That function is misleading, we cannot really translate a graph signal.
I mean, we can compute it (with a multiplication in the spectral domain
by the GFT of a delta), but it won't "translate" the signal (except in
the very special case of shift-invariant graphs, which are graphs for which
the DFT basis vectors are graph Laplacian eigenvectors, e.g., ring, grid).
Try to translate a delta on the ring graph to convince yourself.
We can however localize a graph kernel. See Filter.localize().