nngraph: fix symmetrization
Symmetrizing with the average was setting the distance between v1 and v2
as the average between 0 and the true distance if v1 was the k-NN of v2
but v2 was not in the k-NN of v1.

Moreover, symmetrizing before taking the mean assures that every
distance is counted twice (only some would be counted twice otherwise).